+++ libsolidity/interface/ABI.cpp
@@ -101,7 +101,7 @@ Json::Value ABI::generate(ContractDefinition const& _contractDef)
 			method["stateMutability"] = stateMutabilityToString(externalFunctionType->stateMutability());
 			abi.emplace(std::move(method));
 		}
-	for (auto const& it: _contractDef.definedInterfaceEvents())
+	for (auto const& it: _contractDef.interfaceEvents())
 	{
 		Json::Value event{Json::objectValue};
 		event["type"] = "event";